Migrate codebase to TypeScript
Also changed linter to biome
