+++ package.json
@@ -1,66 +1,74 @@
 {
-  "name": "svelte-spa-router",
-  "version": "5.0.1",
-  "description": "Router for SPAs using Svelte 5",
-  "main": "Router.svelte",
-  "svelte": "Router.svelte",
-  "types": "Router.d.ts",
-  "type": "module",
-  "exports": {
-    "./active": {
-      "types": "./active.d.ts",
-      "svelte": "./active.svelte.js"
+    "name": "svelte-spa-router",
+    "version": "5.0.1",
+    "description": "Router for SPAs using Svelte 5",
+    "type": "module",
+    "files": [
+        "dist"
+    ],
+    "exports": {
+        "./active": {
+            "types": "./dist/active.svelte.d.ts",
+            "svelte": "./dist/active.svelte.js"
+        },
+        "./wrap": {
+            "types": "./dist/wrap.d.ts",
+            "import": "./dist/wrap.js"
+        },
+        ".": {
+            "types": "./dist/Router.svelte.d.ts",
+            "svelte": "./dist/Router.svelte"
+        }
     },
-    "./wrap": {
-      "types": "./wrap.d.ts",
-      "import": "./wrap.js"
+    "packageManager": "pnpm@10.4.0",
+    "scripts": {
+        "build": "svelte-package --input src --output dist",
+        "check": "svelte-check --tsconfig ./tsconfig.json",
+        "build-test-app": "pnpm --dir test/app build",
+        "lint": "biome check .",
+        "lint:fix": "biome check --write .",
+        "format": "biome format --write .",
+        "pretest": "pnpm run build && pnpm --dir test/app install && pnpm run build-test-app",
+        "test": "playwright test",
+        "test:e2e": "playwright test",
+        "prepack": "pnpm run build",
+        "changeset:version": "changeset version && git add --all",
+        "changeset:publish": "changeset publish",
+        "changeset": "changeset"
     },
-    ".": {
-      "types": "./Router.d.ts",
-      "svelte": "./Router.svelte"
+    "repository": {
+        "type": "git",
+        "url": "git+https://github.com/ItalyPaleAle/svelte-spa-router.git"
+    },
+    "keywords": [
+        "router",
+        "svelte",
+        "svelte3",
+        "svelte4",
+        "svelte5",
+        "spa"
+    ],
+    "author": "Alessandro Segala (@ItalyPaleAle)",
+    "license": "MIT",
+    "funding": "https://github.com/sponsors/ItalyPaleAle",
+    "bugs": {
+        "url": "https://github.com/ItalyPaleAle/svelte-spa-router/issues"
+    },
+    "homepage": "https://github.com/ItalyPaleAle/svelte-spa-router#readme",
+    "dependencies": {
+        "regexparam": "2.0.2"
+    },
+    "peerDependencies": {
+        "svelte": "^5.0.0"
+    },
+    "devDependencies": {
+        "@biomejs/biome": "^2.4.12",
+        "@changesets/cli": "^2.29.8",
+        "@playwright/test": "^1.59.1",
+        "@sveltejs/package": "^2.5.7",
+        "@tsconfig/svelte": "^5.0.8",
+        "svelte": "^5.0.0",
+        "svelte-check": "^4.4.6",
+        "typescript": "^5.9.3"
     }
-  },
-  "packageManager": "pnpm@10.4.0",
-  "scripts": {
-    "build-test-app": "pnpm --dir test/app build",
-    "eslint": "npx eslint .",
-    "lint": "npm run eslint",
-    "test": "playwright test",
-    "test:e2e": "playwright test",
-    "changeset:version": "changeset version && git add --all",
-    "changeset:publish": "changeset publish",
-    "changeset": "changeset"
-  },
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/ItalyPaleAle/svelte-spa-router.git"
-  },
-  "keywords": [
-    "router",
-    "svelte",
-    "svelte3",
-    "svelte4",
-    "svelte5",
-    "spa"
-  ],
-  "author": "Alessandro Segala (@ItalyPaleAle)",
-  "license": "MIT",
-  "funding": "https://github.com/sponsors/ItalyPaleAle",
-  "bugs": {
-    "url": "https://github.com/ItalyPaleAle/svelte-spa-router/issues"
-  },
-  "homepage": "https://github.com/ItalyPaleAle/svelte-spa-router#readme",
-  "dependencies": {
-    "regexparam": "2.0.2"
-  },
-  "devDependencies": {
-    "@changesets/cli": "^2.29.8",
-    "@eslint/js": "^9.0.0",
-    "@playwright/test": "^1.59.1",
-    "eslint": "^9.0.0",
-    "eslint-plugin-html": "^7.1.0",
-    "eslint-plugin-svelte": "^3.15.0",
-    "globals": "^16.0.0",
-    "svelte": "^5.0.0"
-  }
 }
