+++ playwright.config.js
@@ -10,7 +10,7 @@ export default defineConfig({
     forbidOnly: !!process.env.CI,
     retries: process.env.CI ? 2 : 0,
     workers: process.env.CI ? 2 : undefined,
-    reporter: process.env.CI ? [ ['github'], ['html'] ] : 'list',
+    reporter: process.env.CI ? [['github'], ['html']] : 'list',
     use: {
         baseURL,
         trace: 'on-first-retry'
